refactor: 💡 adding domain directory
